requirements.txt
bcrypt==4.1.2
blinker==1.6.2
cffi==1.16.0
click==8.1.7
Flask==3.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mistune==3.0.2
mysql-connector-python==8.2.0
protobuf==4.21.12
pycparser==2.21
python-dotenv==1.0.0
watchdog==3.0.0
Werkzeug==3.0.0
